refactor(agent): split storage capability internals
